#27 Move errors to models/erros and error-messages to i18n
