+++ server/src/logic/middleware/auth-validator.middleware.ts
@@ -2,7 +2,7 @@ import * as jwt from "jsonwebtoken";
 import { Request, Response, NextFunction } from "express";
 import { ApiError } from "./../../models/index";
 import { Config } from "./../../config";
-import { ERRORS } from "./../../errors"
+import { ERRORS } from "./../../models/index"
 
 
 export class AuthValidator {
