+++ server/src/routes/api/v1/logs/log.routes.ts
@@ -16,7 +16,7 @@ export class LogRoute {
       (req: Request, res: Response, next: NextFunction) => {
         this._authValidator.checkValidToken(req)
           .then(() => next())
-          .catch((err: any) => res.status(err.statusCode).json(err.forReponse()));
+          .catch((err: any) => res.status(err.statusCode).json(err.forResponse()));
       }, 
       (req: Request, res: Response, next: NextFunction) => {
